Resolve conflicts from the origin/main update, keeping the PR's intent
(oneAPI TBB replaces all OpenMP parallelism) while preserving the features
main added in the meantime:

- ReductionCPU: keep main's two-pass / parallel-dim split for arg-reduction,
  converted to TBB. Use parallel_deterministic_reduce for all floating point
  reductions so results no longer depend on thread count (this fixes the
  test_get_surface_area and test_color_map failures reported on the PR).
- registration/Feature: keep main's SPFH/FPFH subset-index filtering.
- PointCloudSegmentation: keep main's pre-generated RANSAC samples.
- ColorMapUtils: keep main's utility::optional -> std::optional migration.
- VoxelBlockGridImpl: keep main's SYCL atomic min/max paths.

Complete the OpenMP removal so Open3D no longer has an OpenMP dependency:

- Convert the last OpenMP loops (Smoothing.h, PointCloudSmoothing.cpp,
  registration/Feature.cpp) to tbb::parallel_for.
- Remove the WITH_OPENMP option, the OpenMP find_package, the libomp
  packaging for macOS wheels, and the WITH_OPENMP build-config entry.
- Remove utility::OMPProgressBar (superseded by TBBProgressBar) and
  utility::GetThreadNum() / InParallel(), which are OpenMP-only.
- SurfaceReconstructionPoisson: always use PoissonRecon's std::thread pool.
- Replace examples/cpp/OpenMP.cpp with examples/cpp/Parallelism.cpp, which
  demonstrates parallel_for, parallel_reduce and task_arena.

Add a replacement for OMP_NUM_THREADS, which TBB has no equivalent of:
utility::SetMaxThreads() (tbb::global_control), exposed to Python as
open3d.utility.set_max_threads() / get_max_threads(). The reconstruction
system examples use it instead of setting OMP_NUM_THREADS.

Fix bugs found while resolving and reviewing the conversion:

- Registration.cpp: atomic_min's compare-exchange loop exited on contention
  instead of retrying, so the RANSAC iteration bound could be lost.
- Registration.cpp: RANSACCorrespondenceReduction's splitting constructor did
  not size ransac_corres, so split bodies wrote out of bounds.
- LegacyReconstructionUtil.h: PointCloudIntegrator never accumulated into
  fragment, so IntegrateFragmentRGBD produced an empty fragment.
- UniformTSDFVolume::ExtractVoxelGrid wrote into a std::unordered_map from a
  parallel_for; collect into a concurrent_vector and insert serially.
- ColorMapUtils: inverted empty check when averaging invisible-vertex colors.
- ProgressBar::operator+=(n) advanced by 1 instead of n.
- EstimateMaxThreads() ignored tbb::global_control limits.
- Fix '#defined' typos that would break CUDA/SYCL builds of
  VoxelBlockGridImpl.h and FillInLinearSystemImpl.h.
- Consolidate two copies of a double atomic-add helper (whose feature-test
  guard was malformed) into utility::AtomicAdd.
- VoxelBlockGridImpl.h: correct an off-by-one in the ExtractTriangleMesh
  assertion added in #7360; 0 is a valid linear index, and the always-on
  device assert made CUDA tests flaky.

Require oneTBB >= 2021.4.0 (for tbb/collaborative_call_once.h) and link TBB
as a public dependency, since TBB types appear in public Open3D headers.

+++ cpp/open3d/visualization/rendering/RendererHandle.cpp
@@ -1,7 +1,7 @@
 // ----------------------------------------------------------------------------
 // -                        Open3D: www.open3d.org                            -
 // ----------------------------------------------------------------------------
-// Copyright (c) 2018-2023 www.open3d.org
+// Copyright (c) 2018-2024 www.open3d.org
 // SPDX-License-Identifier: MIT
 // ----------------------------------------------------------------------------
 
@@ -13,7 +13,7 @@ namespace open3d {
 namespace visualization {
 namespace rendering {
 
-std::array<std::uint16_t, static_cast<size_t>(EntityType::Count)>
+std::array<std::uint32_t, static_cast<size_t>(EntityType::Count)>
         REHandle_abstract::uid_table;
 
 std::ostream& operator<<(std::ostream& os, const REHandle_abstract& uid) {
